fix: harden CORS defaults, validate route ids, add extension_months int guard
1. CORS (server/index.ts): default origin now covers both Vite dev-server
   hostnames (localhost:5173 and 127.0.0.1:5173 as an array); startup guard
   exits with a fatal error when CORS_ORIGIN='*' to prevent wildcard
   misconfiguration in production.

2. Route id validation (server/routes.ts): PATCH /projects/:id and
   DELETE /staffing/:id now return 400 for any non-positive or non-integer
   id (e.g. 'abc', '1.5', '-1', '0') before touching the DB.
   PATCH returns 404 when valid fields are supplied for a non-existent id.
   Empty-body PATCH remains a 200 no-op (response now includes updated:0).

3. updateProject return type (server/db.ts): function now returns
   Database.RunResult so the route can inspect changes count for 404
   detection; a no-op (empty fields) returns a synthetic {changes:0}.

4. extension_months int guard (server/engine/validation.ts): adds .int()
   to match count and frequency_months — fractional values are now rejected
   by the Zod schema.

+++ server/engine/validation.ts
@@ -61,7 +61,7 @@ export const scenarioOperationSchema: z.ZodType<ScenarioOperation> = z.object({
   rate_changes: z.array(rateChangeSchema).optional(),
   hours_changes: z.array(hoursChangeSchema).optional(),
   new_end_date: z.string().optional(),
-  extension_months: z.number().positive().optional(),
+  extension_months: z.number().int().positive().optional(),
   additional_costs: z.array(additionalCostSchema).optional(),
   sub_operations: z.array(z.lazy(() => scenarioOperationSchema)).optional(),
   _fallback: z.boolean().optional(),
